feat: Refactor cover letter and resume components for improved structure and organization
- Updated import paths for components in both cover letter and resume templates to enhance modularity.
- Introduced new components for cover letter sections including Header, DateLine, Body, Title, and Signature, improving clarity and maintainability.
- Added new components for resume sections including Header, Contact, Skills, Languages, Education, Experience, and Title, streamlining the overall structure.
- Enhanced document properties configuration for better user experience and consistency across templates.
