+++ src/templates/modern/resume/components/Education.tsx
