+++ src/templates/modern/resume/components/List.tsx
